+++ app/src/main/java/com/ivanmorgillo/corsoandroid/teama/extension/KotlinExt.kt
@@ -1,4 +1,4 @@
-package com.ivanmorgillo.corsoandroid.teama
+package com.ivanmorgillo.corsoandroid.teama.extension
 
 val <T> T.exhaustive: T
     get() = this
