+++ app/src/main/java/com/ivanmorgillo/corsoandroid/teama/recipe/MainViewModel.kt
@@ -9,6 +9,7 @@ import com.ivanmorgillo.corsoandroid.teama.MainScreenStates.Content
 import com.ivanmorgillo.corsoandroid.teama.MainScreenStates.Error
 import com.ivanmorgillo.corsoandroid.teama.MainScreenStates.Loading
 import com.ivanmorgillo.corsoandroid.teama.crashlytics.SingleLiveEvent
+import com.ivanmorgillo.corsoandroid.teama.extension.exhaustive
 import com.ivanmorgillo.corsoandroid.teama.network.LoadRecipeError
 import com.ivanmorgillo.corsoandroid.teama.network.LoadRecipeResult.Failure
 import com.ivanmorgillo.corsoandroid.teama.network.LoadRecipeResult.Success
